obdlink/requirements.txt
obd==0.7.1
paho-mqtt==1.6.1
pint>=0.21,<0.23
